+++ reductstore/src/api/http.rs
@@ -893,7 +893,6 @@ pub(crate) mod tests {
                     dst_token: None,
                     entries: vec![],
                     dst_prefix: String::new(),
-                    each_n: None,
                     when: None,
                     mode: ReplicationMode::Enabled,
                     compression: ReplicationCompression::None,